update generateChunnkSources to single location
