+++ frontend/src/components/Sidebar/ActiveWorkspaces/index.jsx
@@ -161,8 +161,8 @@ export default function ActiveWorkspaces() {
                               isInWorkspaceSettings && workspace.slug === slug
                                 ? "#46C8FF"
                                 : settingHover[workspace.id]
-                                ? "#FFFFFF"
-                                : "#A7A8A9"
+                                  ? "#FFFFFF"
+                                  : "#A7A8A9"
                             }
                             weight="bold"
                             className="h-[20px] w-[20px]"
